requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
pillow==10.2.0
pycparser==2.21
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
